backend/requirements.txt
asgiref==3.2.3
DAWG-Python==0.7.2
Django==3.0.2
django-cors-headers==3.2.1
django-filter==2.2.0
django-model-utils==4.0.0
djangorestframework==3.12.4
djangorestframework-jsonapi==4.2.1
docopt==0.6.2
docx==0.2.4
psycopg2-binary==2.8.4
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.404381.4453942
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
Pillow==6.2.1
pyexcel-xlsx==0.6.0
drf-nested-routers==0.93.4
requests==2.26.0